refactor(web): decompose AssistantShell into RootLayout + ChatLayout
Extract the monolithic AssistantShell into proper React Router layout
routes following the *Layout naming convention:

- RootLayout (components/layout/root-layout.tsx): App-level concerns —
  safe-area insets, iOS visual-viewport keyboard tracking. Pathless
  layout route wrapping all child layouts.

- ChatLayout (domains/chat/chat-layout.tsx): Chat-specific concerns —
  sidebar rail, mobile drawer with focus trap, keyboard shortcuts
  (Ctrl+\, Ctrl+K, Ctrl+[/]), header with back/forward navigation.
  Absorbs the history tracking logic from App.tsx.

- SideMenu moved to domains/chat/ — it's chat-domain navigation,
  not cross-domain shared UI.

This enables SettingsLayout to be added as a sibling layout route to
ChatLayout (both children of RootLayout), matching the platform's
architecture where SettingsShell and AssistantShell are siblings.

References:
- React Router layout routes: https://reactrouter.com/start/data/routing
- React Router nested routes: https://reactrouter.com/start/data/routing#nested-routes
- Visual Viewport API: https://developer.mozilla.org/en-US/docs/Web/API/Visual_Viewport_API
